This is a very basic implementation of a Sound Blaster driver for ReactOS. At present, it's capable of playing a buffer of silence when opened, but that's about all.
The MPU401 driver would normally be part of SNDBLST, but I decided to keep them separate, for now at least anyway. IIRC, this should play a MIDI note on startup.

Both drivers need a lot of work, but I'm committing what I've done so far, so other people can work on them as well as me :)