Revert hpoussin's commit in r29960 to unbreak build.
Most of *our* stuff has separated host/target headers now, but the Wine-shared components still include headers from include/reactos/wine.
And these headers (like "winnt.h" in that directory) include our headers from include/psdk themselves, which are meant to be used for target components, but not for host tools.

This needs to be fixed first, before we can remove this hack.