Implement a new way to build only the required parts of the reactos codebase partly inspired by bug 2568.
Introduced the concept of module "family" and "buildfamilies"
Any module can belong to an N number of families for example the 'calc' win32 gui module:

<module name="calc" type="win32gui" installbase="system32" installname="calc.exe">
	(..)
	<family>applications</family>
	<family>guiapplications</family>
	(..)

and the 'solitaire' win32 gui module:

<module name="sol" type="win32gui" installbase="system32" installname="sol.exe" unicode="no" allowwarnings="true">
	(..)
	<family>applications</family>
	<family>guiapplications</family>
	<family>games</family>
	(..)

now is possible to type on the rosbe console:

>make applications (calc and sol and all its dependencies will be build)
>make applications_clean (calc and sol clean target will be executed)
>make games (sol and any other member of the games family will be build)

It works in conjunction with the new \buildfamilies.rbuild. In a properly tagged codebase the developer has the option to build only the parts of the operating system he is currently working on without having to clean and recompile the whole codebase every time. All required dependencies will be also automatically compiled.