Sync to wine-0.9.60: - Rob Shearman <rob@codeweavers.com> Fri, 11 Apr 2008 widl: Support non-default calling conventions for object methods.
- Rob Shearman <rob@codeweavers.com> Fri, 11 Apr 2008
widl: Support non-default calling conventions for non-object functions.

- Rob Shearman <rob@codeweavers.com> Mon, 14 Apr 2008
widl: Check that attributes applied to interfaces, functions and arguments are applicable and issue an error if not.

- Rob Shearman <rob@codeweavers.com> Mon, 14 Apr 2008
widl: Check that attributes applied to typedefs and fields are applicable and issue an error otherwise.

- Rob Shearman <rob@codeweavers.com> Mon, 14 Apr 2008
widl: Make the attrs parameter passed to start_typelib const.

- Rob Shearman <rob@codeweavers.com> Mon, 14 Apr 2008
widl: The odl attribute is valid for dispinterfaces.

- Rob Shearman <rob@codeweavers.com> Mon, 14 Apr 2008
widl: Check that attributes are applicable for libraries,  modules, dispinterfaces and coclasses and otherwise issue an error.

- Rob Shearman <rob@codeweavers.com> Mon, 14 Apr 2008
widl: Output a warning if duplicate attributes are specified.

- Rob Shearman <rob@codeweavers.com> Mon, 14 Apr 2008
widl: Add support for the broadcast and idempotent operation attributes.

- Rob Shearman <rob@codeweavers.com> Mon, 14 Apr 2008
widl: Write out TYPEFLAG, FUNCFLAG and VARFLAG flags for all attribute we currently parse.
Add FIXMEs for the remaining ones and remove the warning for unimplemented attribute types since we now support all the attributes that we parse.

- Rob Shearman <rob@codeweavers.com> Mon, 14 Apr 2008
widl: Add functions to print an error or warning message using location information to enable these to be printed accurately when a check is done after an element is parsed.
Add location information to variable automatically, since this is will be useful for type checking of arguments.

- Rob Shearman <rob@codeweavers.com> Mon, 14 Apr 2008
widl: Add checking to arguments of non-local functions.
Check that out arguments have enough levels of indirection and that they don't derive from void * or a function pointer.

- Rob Shearman <rob@codeweavers.com> Tue, 15 Apr 2008
widl: Change alls calls to error in the parser to error_loc so that line number information is printed.