[IP] - Rewrite most of the lower TCP layer code to work with lwIP [LWIP] - Write all of the implementation-specific code and headers [TCPIP] - Add assertions in the object reference functions - Remove the completion queue hack
This is still isn't totally finished yet. There are still two main issues that need to be overcome: listen/accept don't work at all and the performance is horrible.
I hope that once I tune some of the options, I can get it to perform reasonably well. If any curious testers want to try this out, be my guest but realize that not
a lot of apps will work yet. Suggestions are welcome! :)